countdown.py
I like this Python project concept! Here, we've developed a countdown timer that requests a user-inputted number of seconds, counts down from that number second by second, and then displays a message.

We have set pauses of 1 second using the Python time module's.sleep() function. To create the countdown display, we use this in conjunction with some clever string formatting.